+++ MarkdownMonster/Editor/table-editor.js
@@ -1,0 +1,16 @@
+﻿var app = {
+    columnCount: 3,
+    rowCount: 3,
+    columnData: [
+        [
+            "R1 Column 1 text",
+            "R1 Column 2 text",
+            "R1 Column 3 text"
+        ],
+        [
+            "R1 Column 1 text",
+            "R1 Column 2 text",
+            "R1 Column 3 text"
+        ]
+    ]
+}